Completed Tugas 6

+++ todolist/views.py
@@ -1,3 +1,5 @@
+import json
+import re
 from django.shortcuts import render
 
 # Create task
@@ -18,18 +20,25 @@ from .models import Task
 # Restriksi halaman todolist
 from django.contrib.auth.decorators import login_required
 
+# Data Delivery
+from django.http import HttpResponse, JsonResponse
+from django.core import serializers
+
 
 # Show todolist
 @login_required(login_url='/todolist/login/')
 def show_todolist(request):
+    form = forms.CreateTask()
     tasks = Task.objects.filter(user=request.user)
     context = {
         'user' : request.user,
         'tasks' : tasks,
+        'form' : form,
     }
     return render(request, 'todolist.html', context)
 
 # Create task
+"""
 @login_required(login_url='/todolist/login/')
 def create_task(request):
     form = forms.CreateTask()
@@ -44,6 +53,7 @@ def create_task(request):
             
     context = {'form':form}
     return render(request, 'create_task.html', context)
+"""
 
 # Register user
 def register(request):
@@ -83,10 +93,37 @@ def change_status(request, id):
     task = Task.objects.get(user=request.user, pk=id)
     task.is_finished = not task.is_finished
     task.save()
-    return redirect('todolist:show_todolist')
+
+    task = Task.objects.filter(user=request.user, pk=id)
+    return HttpResponse(serializers.serialize("json", task), content_type="application/json")
 
 # Delete task
 def delete_task(request, id):
-    task = Task.objects.get(user=request.user, pk=id)
-    task.delete()
-    return redirect('todolist:show_todolist')
+    if request.method == "DELETE":
+        task = Task.objects.get(user=request.user, pk=id)
+        task.delete()
+
+        task = Task.objects.filter(user=request.user)
+        return HttpResponse(serializers.serialize("json", task), content_type="application/json")
+    return HttpResponse('')
+
+
+# Menampilkan data dalam bentuk JSON
+def show_json(request):
+    task = Task.objects.filter(user=request.user)
+    return HttpResponse(serializers.serialize("json", task), content_type="application/json")
+
+# Add task
+def add_task(request):
+    if request.method == "POST":
+        form = forms.CreateTask()
+        instance = form.save(commit=False)
+        instance.user = request.user
+        instance.title = request.POST.get('title')
+        instance.description = request.POST.get('description')
+        instance.save()
+        id = instance.pk
+
+        task = Task.objects.filter(user=request.user, pk=id)
+        return HttpResponse(serializers.serialize("json", task), content_type="application/json")
+    return HttpResponse('')
